Added a fullscreen WebGL template.
Map Generator :
fixed the 2d noise issue (it was generating pseudo non-coherent noise because it's ln lat width was too big).
added plenty of new features :
a random seed selector
a seamless texture option
a noise start + width position field
a custom gradient selection.
The font of all the texts are now bigger so we can read it.

Solar system generator :
Remade all the planet profiles and got rid of all the legacy ones (since they aren't suitable anymore for a RGB Splatmap based texturing).
the sun have it's texture again.
Commented some of the PlanetDisplayer in order to be able to compile, will be placed in an Editor folder soon.